refactor: Split aggregate/expression.rs into focused submodules
Reorganize the 1405-line expression.rs into a modular structure:

- expression/mod.rs: Public API, extract_aggregates, re-exports
- expression/evaluator.rs: Simple expression evaluation (eval_simple_expr)
- expression/vectorized.rs: Vectorized binary aggregate operations
- expression/simd.rs: SIMD-accelerated computation via Arrow
- expression/batch.rs: Batch-native processing on ColumnarBatch
- expression/tests.rs: All tests preserved

Benefits:
- Each file focuses on one concern (~50-400 lines each)
- Easier to unit test individual components
- Faster to navigate and understand
- Potential for better incremental compilation

All existing tests pass. No performance regression expected as
this is a pure code organization change with no logic modifications.

Closes #2701
