--- crates/vibesql-executor/src/select/columnar/aggregate/expression.rs
@@ -1,1405 +1,0 @@
-//! Expression aggregates - aggregating over expressions rather than simple columns
-//!
-//! This module handles aggregates over complex expressions like SUM(a * b),
-//! where we need to evaluate the expression for each row before aggregating.
-//!
-//! For large datasets (>= 100 rows), this module automatically uses SIMD-accelerated
-//! evaluation via Apache Arrow, providing 4-8x performance improvement.
-//!
-//! ## Batch-Native Path
-//!
-//! The `compute_batch_expression_aggregate` function provides SIMD-accelerated
-//! expression evaluation directly on `ColumnarBatch` without converting to rows.
-//! This is ~20-30% faster than the row-based path for large batches.
-
-use crate::errors::ExecutorError;
-use crate::schema::CombinedSchema;
-use vibesql_ast::Expression;
-use vibesql_storage::Row;
-use vibesql_types::SqlValue;
-
-use super::functions::compare_for_min_max;
-use super::{AggregateOp, AggregateSource, AggregateSpec};
-use super::super::scan::ColumnarScan;
-use super::super::batch::{ColumnarBatch, ColumnArray};
-
-/// Threshold for using SIMD acceleration (same as vectorized filter threshold)
-const SIMD_THRESHOLD: usize = 100;
-
-/// Evaluate a simple arithmetic expression for a single row
-///
-/// This is a lightweight evaluator for the subset of expressions we support
-/// in columnar aggregates (column references and binary operations).
-pub(super) fn eval_simple_expr(
-    expr: &Expression,
-    row: &Row,
-    schema: &CombinedSchema,
-) -> Result<SqlValue, ExecutorError> {
-    match expr {
-        Expression::ColumnRef { table, column } => {
-            let col_idx = schema.get_column_index(table.as_deref(), column)
-                .ok_or_else(|| ExecutorError::UnsupportedExpression(
-                    format!("Column not found: {}", column)
-                ))?;
-            Ok(row.get(col_idx).cloned().unwrap_or(SqlValue::Null))
-        }
-        Expression::Literal(val) => Ok(val.clone()),
-        Expression::BinaryOp { left, op, right } => {
-            let left_val = eval_simple_expr(left, row, schema)?;
-            let right_val = eval_simple_expr(right, row, schema)?;
-
-            // Use the evaluator's operators module
-            use crate::evaluator::operators::OperatorRegistry;
-            OperatorRegistry::eval_binary_op(&left_val, op, &right_val, vibesql_types::SqlMode::default())
-        }
-        _ => Err(ExecutorError::UnsupportedExpression(
-            "Complex expressions not supported in columnar aggregates".to_string()
-        )),
-    }
-}
-
-/// Try vectorized evaluation for simple binary operations (e.g., col_a * col_b)
-///
-/// Returns Some(result) if the expression can be vectorized (simple column * column),
-/// or None if we need to fall back to row-by-row evaluation.
-fn try_vectorized_binary_aggregate(
-    rows: &[Row],
-    expr: &Expression,
-    op: AggregateOp,
-    filter_bitmap: Option<&[bool]>,
-    schema: &CombinedSchema,
-) -> Result<Option<SqlValue>, ExecutorError> {
-    // Only optimize SUM and AVG for now
-    if !matches!(op, AggregateOp::Sum | AggregateOp::Avg) {
-        return Ok(None);
-    }
-
-    // Check if expression is a simple binary multiply: col_a * col_b
-    if let Expression::BinaryOp { left, op: bin_op, right } = expr {
-        use vibesql_ast::BinaryOperator;
-
-        // Only handle multiplication for now (most common in TPC-H)
-        if *bin_op != BinaryOperator::Multiply {
-            return Ok(None);
-        }
-
-        // Both operands must be simple column references
-        let (left_col, right_col) = match (left.as_ref(), right.as_ref()) {
-            (
-                Expression::ColumnRef { table: t1, column: c1 },
-                Expression::ColumnRef { table: t2, column: c2 }
-            ) => {
-                let idx1 = schema.get_column_index(t1.as_deref(), c1)
-                    .ok_or_else(|| ExecutorError::UnsupportedExpression(
-                        format!("Column not found: {}", c1)
-                    ))?;
-                let idx2 = schema.get_column_index(t2.as_deref(), c2)
-                    .ok_or_else(|| ExecutorError::UnsupportedExpression(
-                        format!("Column not found: {}", c2)
-                    ))?;
-                (idx1, idx2)
-            }
-            _ => return Ok(None), // Not simple col * col pattern
-        };
-
-        // Vectorized path: extract both columns, multiply, sum
-        let scan = ColumnarScan::new(rows);
-        let mut sum = 0.0;
-        let mut count = 0;
-
-        // Batch processing: accumulate products in batches for better cache locality
-        const BATCH_SIZE: usize = 1024;
-        let mut batch_products = Vec::with_capacity(BATCH_SIZE);
-
-        for row_idx in 0..rows.len() {
-            // Check filter
-            if let Some(bitmap) = filter_bitmap {
-                if !bitmap.get(row_idx).copied().unwrap_or(false) {
-                    continue;
-                }
-            }
-
-            // Get values from both columns
-            let val1 = scan.row(row_idx)
-                .and_then(|row| row.get(left_col))
-                .unwrap_or(&SqlValue::Null);
-            let val2 = scan.row(row_idx)
-                .and_then(|row| row.get(right_col))
-                .unwrap_or(&SqlValue::Null);
-
-            // Convert to f64 and multiply
-            if let (Some(v1), Some(v2)) = (sql_value_to_f64(val1), sql_value_to_f64(val2)) {
-                batch_products.push(v1 * v2);
-                count += 1;
-
-                // Process batch when full
-                if batch_products.len() >= BATCH_SIZE {
-                    sum += batch_products.iter().sum::<f64>();
-                    batch_products.clear();
-                }
-            }
-        }
-
-        // Process remaining batch
-        if !batch_products.is_empty() {
-            sum += batch_products.iter().sum::<f64>();
-        }
-
-        let result = if count > 0 {
-            match op {
-                AggregateOp::Sum => SqlValue::Double(sum),
-                AggregateOp::Avg => SqlValue::Double(sum / count as f64),
-                _ => unreachable!(),
-            }
-        } else {
-            SqlValue::Null
-        };
-
-        return Ok(Some(result));
-    }
-
-    Ok(None)
-}
-
-/// Convert SqlValue to f64 for arithmetic operations
-fn sql_value_to_f64(val: &SqlValue) -> Option<f64> {
-    match val {
-        SqlValue::Integer(v) => Some(*v as f64),
-        SqlValue::Bigint(v) => Some(*v as f64),
-        SqlValue::Smallint(v) => Some(*v as f64),
-        SqlValue::Float(v) => Some(*v as f64),
-        SqlValue::Double(v) => Some(*v),
-        SqlValue::Numeric(v) => Some(*v),
-        SqlValue::Null => None,
-        _ => None,
-    }
-}
-
-/// Compute an aggregate over an expression (e.g., SUM(a * b))
-///
-/// Evaluates the expression for each row, then aggregates the results.
-/// For large datasets (>= SIMD_THRESHOLD rows), uses SIMD-accelerated evaluation.
-pub(super) fn compute_expression_aggregate(
-    rows: &[Row],
-    expr: &Expression,
-    op: AggregateOp,
-    filter_bitmap: Option<&[bool]>,
-    schema: &CombinedSchema,
-) -> Result<SqlValue, ExecutorError> {
-    // Try main branch's vectorized path first for simple binary operations
-    // This is optimized for column × column multiplication with optional filtering
-    if let Some(result) = try_vectorized_binary_aggregate(rows, expr, op, filter_bitmap, schema)? {
-        return Ok(result);
-    }
-
-    // Try SIMD path for large datasets (more general than vectorized binary)
-    // Only when no filter bitmap (vectorized binary handles filtered case)
-    if rows.len() >= SIMD_THRESHOLD && filter_bitmap.is_none() {
-        if let Ok(result) = try_simd_aggregate(rows, expr, op, schema) {
-            return Ok(result);
-        }
-        // Fall through to scalar path if SIMD fails
-    }
-
-    // Scalar path (for small datasets, complex expressions, or when SIMD not applicable)
-    match op {
-        AggregateOp::Sum => {
-            let mut int_sum: i64 = 0;
-            let mut float_sum = 0.0;
-            let mut count = 0;
-            let mut has_float = false;
-
-            for (row_idx, row) in rows.iter().enumerate() {
-                // Check filter bitmap
-                if let Some(bitmap) = filter_bitmap {
-                    if !bitmap.get(row_idx).copied().unwrap_or(false) {
-                        continue;
-                    }
-                }
-
-                // Evaluate expression for this row
-                let value = eval_simple_expr(expr, row, schema)?;
-
-                // Add to sum
-                if !matches!(value, SqlValue::Null) {
-                    match value {
-                        SqlValue::Integer(v) => {
-                            if has_float {
-                                float_sum += v as f64;
-                            } else {
-                                int_sum += v;
-                            }
-                        }
-                        SqlValue::Bigint(v) => {
-                            if has_float {
-                                float_sum += v as f64;
-                            } else {
-                                int_sum += v;
-                            }
-                        }
-                        SqlValue::Smallint(v) => {
-                            if has_float {
-                                float_sum += v as f64;
-                            } else {
-                                int_sum += v as i64;
-                            }
-                        }
-                        SqlValue::Float(v) => {
-                            if !has_float {
-                                // Convert accumulated integer sum to float
-                                float_sum = int_sum as f64;
-                                has_float = true;
-                            }
-                            float_sum += v as f64;
-                        }
-                        SqlValue::Double(v) => {
-                            if !has_float {
-                                // Convert accumulated integer sum to float
-                                float_sum = int_sum as f64;
-                                has_float = true;
-                            }
-                            float_sum += v;
-                        }
-                        SqlValue::Numeric(v) => {
-                            if !has_float {
-                                // Convert accumulated integer sum to float
-                                float_sum = int_sum as f64;
-                                has_float = true;
-                            }
-                            float_sum += v;
-                        }
-                        SqlValue::Null => {}, // Already checked above
-                        _ => {
-                            return Err(ExecutorError::UnsupportedExpression(
-                                format!("Cannot compute SUM on non-numeric value: {:?}", value)
-                            ))
-                        }
-                    }
-                    count += 1;
-                }
-            }
-
-            Ok(if count > 0 {
-                if has_float {
-                    SqlValue::Double(float_sum)
-                } else {
-                    SqlValue::Integer(int_sum)
-                }
-            } else {
-                SqlValue::Null
-            })
-        }
-        AggregateOp::Count => {
-            // COUNT of expression counts non-NULL results
-            let mut count = 0;
-            for (row_idx, row) in rows.iter().enumerate() {
-                if let Some(bitmap) = filter_bitmap {
-                    if !bitmap.get(row_idx).copied().unwrap_or(false) {
-                        continue;
-                    }
-                }
-                let value = eval_simple_expr(expr, row, schema)?;
-                if !matches!(value, SqlValue::Null) {
-                    count += 1;
-                }
-            }
-            Ok(SqlValue::Integer(count))
-        }
-        AggregateOp::Avg => {
-            // AVG(expr) = SUM(expr) / COUNT(expr)
-            let sum_result = compute_expression_aggregate(rows, expr, AggregateOp::Sum, filter_bitmap, schema)?;
-            let count_result = compute_expression_aggregate(rows, expr, AggregateOp::Count, filter_bitmap, schema)?;
-
-            match (sum_result, count_result) {
-                (SqlValue::Integer(sum), SqlValue::Integer(count)) if count > 0 => {
-                    Ok(SqlValue::Double(sum as f64 / count as f64))
-                }
-                (SqlValue::Double(sum), SqlValue::Integer(count)) if count > 0 => {
-                    Ok(SqlValue::Double(sum / count as f64))
-                }
-                _ => Ok(SqlValue::Null),
-            }
-        }
-        AggregateOp::Min | AggregateOp::Max => {
-            let mut result_value: Option<SqlValue> = None;
-
-            for (row_idx, row) in rows.iter().enumerate() {
-                if let Some(bitmap) = filter_bitmap {
-                    if !bitmap.get(row_idx).copied().unwrap_or(false) {
-                        continue;
-                    }
-                }
-
-                let value = eval_simple_expr(expr, row, schema)?;
-                if !matches!(value, SqlValue::Null) {
-                    result_value = Some(match &result_value {
-                        None => value,
-                        Some(current) => {
-                            let should_update = if op == AggregateOp::Min {
-                                compare_for_min_max(&value, current)
-                            } else {
-                                compare_for_min_max(current, &value)
-                            };
-                            if should_update {
-                                value
-                            } else {
-                                current.clone()
-                            }
-                        }
-                    });
-                }
-            }
-
-            Ok(result_value.unwrap_or(SqlValue::Null))
-        }
-    }
-}
-
-/// Extract aggregate operations from AST expressions
-///
-/// Converts aggregate function expressions to AggregateSpec objects
-/// that can be used with columnar execution.
-///
-/// Currently supports:
-/// - SUM(column) → Column aggregate (fast path)
-/// - SUM(a * b) → Expression aggregate (evaluates expression per row)
-/// - COUNT(*) or COUNT(column) → Column aggregate
-/// - AVG(column) or AVG(expr) → Column/Expression aggregate
-/// - MIN(column) or MIN(expr) → Column/Expression aggregate
-/// - MAX(column) or MAX(expr) → Column/Expression aggregate
-///
-/// Supported expression types:
-/// - Simple column references (fast path)
-/// - Binary operations (+, -, *, /) with column references
-///
-/// Returns None if the expression contains unsupported patterns:
-/// - DISTINCT aggregates
-/// - Multiple arguments
-/// - Complex expressions (subqueries, function calls, etc.)
-/// - Non-aggregate expressions
-///
-/// # Arguments
-///
-/// * `exprs` - The SELECT list expressions
-/// * `schema` - The schema to resolve column names to indices
-///
-/// # Returns
-///
-/// Some(aggregates) if all expressions can be converted to aggregates,
-/// None if any expression is too complex for columnar optimization.
-pub fn extract_aggregates(
-    exprs: &[Expression],
-    schema: &CombinedSchema,
-) -> Option<Vec<AggregateSpec>> {
-    let mut aggregates = Vec::new();
-
-    for expr in exprs.iter() {
-        match expr {
-            Expression::AggregateFunction {
-                name,
-                distinct,
-                args,
-            } => {
-                // DISTINCT not supported for columnar optimization
-                if *distinct {
-                    return None;
-                }
-
-                let op = match name.to_uppercase().as_str() {
-                    "SUM" => AggregateOp::Sum,
-                    "COUNT" => AggregateOp::Count,
-                    "AVG" => AggregateOp::Avg,
-                    "MIN" => AggregateOp::Min,
-                    "MAX" => AggregateOp::Max,
-                    _ => return None, // Unsupported aggregate function
-                };
-
-                // Handle COUNT(*)
-                if op == AggregateOp::Count && args.is_empty() {
-                    // For COUNT(*), use column 0 (the column index is ignored by compute_count)
-                    aggregates.push(AggregateSpec {
-                        op,
-                        source: AggregateSource::Column(0),
-                    });
-                    continue;
-                }
-
-                // Handle COUNT(*) with wildcard argument (Expression::Wildcard or ColumnRef { column: "*" })
-                if op == AggregateOp::Count && args.len() == 1 {
-                    match &args[0] {
-                        Expression::Wildcard => {
-                            aggregates.push(AggregateSpec {
-                                op,
-                                source: AggregateSource::CountStar,
-                            });
-                            continue;
-                        }
-                        Expression::ColumnRef { table: _, column } if column == "*" => {
-                            aggregates.push(AggregateSpec {
-                                op,
-                                source: AggregateSource::CountStar,
-                            });
-                            continue;
-                        }
-                        _ => {}
-                    }
-                }
-
-                // Extract source (column or expression) for other aggregates
-                if args.len() != 1 {
-                    return None; // Multiple arguments not supported
-                }
-
-                let source = match &args[0] {
-                    // Fast path: simple column reference
-                    Expression::ColumnRef { table, column } => {
-                        let column_idx = schema.get_column_index(table.as_deref(), column)?;
-                        AggregateSource::Column(column_idx)
-                    }
-                    // New: support binary operations like a * b
-                    Expression::BinaryOp { .. } => {
-                        // Check if this is a simple binary operation we can handle
-                        if is_simple_arithmetic_expr(&args[0], schema).is_some() {
-                            AggregateSource::Expression(args[0].clone())
-                        } else {
-                            return None; // Complex expression not supported
-                        }
-                    }
-                    _ => return None, // Other expression types not supported
-                };
-
-                aggregates.push(AggregateSpec { op, source });
-            }
-            _ => {
-                return None; // Non-aggregate expressions not supported
-            }
-        }
-    }
-
-    Some(aggregates)
-}
-
-/// Check if an expression is a simple arithmetic expression we can optimize
-///
-/// Returns Some(()) if the expression only contains column references and
-/// arithmetic operations (+, -, *, /), which we can efficiently evaluate.
-/// Returns None if the expression contains unsupported operations.
-fn is_simple_arithmetic_expr(expr: &Expression, schema: &CombinedSchema) -> Option<()> {
-    match expr {
-        Expression::ColumnRef { table, column } => {
-            // Verify column exists
-            schema.get_column_index(table.as_deref(), column)?;
-            Some(())
-        }
-        Expression::Literal(_) => Some(()),
-        Expression::BinaryOp { left, op, right } => {
-            // Only support arithmetic operations
-            use vibesql_ast::BinaryOperator::*;
-            match op {
-                Plus | Minus | Multiply | Divide => {
-                    is_simple_arithmetic_expr(left, schema)?;
-                    is_simple_arithmetic_expr(right, schema)?;
-                    Some(())
-                }
-                _ => None, // Comparison ops, logical ops, etc. not supported
-            }
-        }
-        _ => None, // Function calls, subqueries, etc. not supported
-    }
-}
-
-/// Try to compute aggregate using SIMD-accelerated evaluation
-///
-/// This function converts rows to Arrow RecordBatch, evaluates the expression
-/// using SIMD operations, and aggregates the result.
-///
-/// Returns Ok(value) if SIMD path succeeds, Err(_) if it fails (caller falls back to scalar).
-fn try_simd_aggregate(
-    rows: &[Row],
-    expr: &Expression,
-    op: AggregateOp,
-    schema: &CombinedSchema,
-) -> Result<SqlValue, ExecutorError> {
-    use crate::select::vectorized::{
-        evaluate_arithmetic_simd, rows_to_record_batch,
-    };
-
-    // Extract column names from schema (in order)
-    let mut column_names = vec![String::new(); schema.total_columns];
-    for (start_idx, table_schema) in schema.table_schemas.values() {
-        for (col_idx, col) in table_schema.columns.iter().enumerate() {
-            column_names[start_idx + col_idx] = col.name.clone();
-        }
-    }
-
-    // Convert rows to RecordBatch
-    let batch = rows_to_record_batch(rows, &column_names)
-        .map_err(|_| ExecutorError::Other("Failed to convert to RecordBatch".to_string()))?;
-
-    // Evaluate expression using SIMD
-    let result_array = evaluate_arithmetic_simd(&batch, expr)?;
-
-    // Aggregate the result array
-    match op {
-        AggregateOp::Sum => {
-            // Use Arrow compute kernels for summing
-            use arrow::array::{Float64Array, Int64Array};
-            use arrow::compute::sum;
-
-            match result_array.data_type() {
-                arrow::datatypes::DataType::Int64 => {
-                    let arr = result_array
-                        .as_any()
-                        .downcast_ref::<Int64Array>()
-                        .ok_or_else(|| {
-                            ExecutorError::Other("Failed to downcast Int64Array".to_string())
-                        })?;
-                    let sum_val = sum(arr).ok_or_else(|| {
-                        ExecutorError::Other("SIMD sum returned None".to_string())
-                    })?;
-                    Ok(SqlValue::Integer(sum_val))
-                }
-                arrow::datatypes::DataType::Float64 => {
-                    let arr = result_array
-                        .as_any()
-                        .downcast_ref::<Float64Array>()
-                        .ok_or_else(|| {
-                            ExecutorError::Other("Failed to downcast Float64Array".to_string())
-                        })?;
-                    let sum_val = sum(arr).ok_or_else(|| {
-                        ExecutorError::Other("SIMD sum returned None".to_string())
-                    })?;
-                    Ok(SqlValue::Double(sum_val))
-                }
-                _ => Err(ExecutorError::Other("Unsupported array type for SUM".to_string())),
-            }
-        }
-        AggregateOp::Count => {
-            // Count non-null values
-            let non_null_count = result_array.len() - result_array.null_count();
-            Ok(SqlValue::Integer(non_null_count as i64))
-        }
-        AggregateOp::Avg => {
-            // AVG = SUM / COUNT
-            let sum_result = try_simd_aggregate(rows, expr, AggregateOp::Sum, schema)?;
-            let count_result = try_simd_aggregate(rows, expr, AggregateOp::Count, schema)?;
-
-            match (sum_result, count_result) {
-                (SqlValue::Double(sum), SqlValue::Integer(count)) if count > 0 => {
-                    Ok(SqlValue::Double(sum / count as f64))
-                }
-                (SqlValue::Integer(sum), SqlValue::Integer(count)) if count > 0 => {
-                    Ok(SqlValue::Double(sum as f64 / count as f64))
-                }
-                _ => Ok(SqlValue::Null),
-            }
-        }
-        AggregateOp::Min => {
-            use arrow::array::{Float64Array, Int64Array};
-            use arrow::compute::min;
-
-            match result_array.data_type() {
-                arrow::datatypes::DataType::Int64 => {
-                    let arr = result_array
-                        .as_any()
-                        .downcast_ref::<Int64Array>()
-                        .ok_or_else(|| {
-                            ExecutorError::Other("Failed to downcast Int64Array".to_string())
-                        })?;
-                    let min_val = min(arr).ok_or_else(|| {
-                        ExecutorError::Other("SIMD min returned None".to_string())
-                    })?;
-                    Ok(SqlValue::Integer(min_val))
-                }
-                arrow::datatypes::DataType::Float64 => {
-                    let arr = result_array
-                        .as_any()
-                        .downcast_ref::<Float64Array>()
-                        .ok_or_else(|| {
-                            ExecutorError::Other("Failed to downcast Float64Array".to_string())
-                        })?;
-                    let min_val = min(arr).ok_or_else(|| {
-                        ExecutorError::Other("SIMD min returned None".to_string())
-                    })?;
-                    Ok(SqlValue::Double(min_val))
-                }
-                _ => Err(ExecutorError::Other("Unsupported array type for MIN".to_string())),
-            }
-        }
-        AggregateOp::Max => {
-            use arrow::array::{Float64Array, Int64Array};
-            use arrow::compute::max;
-
-            match result_array.data_type() {
-                arrow::datatypes::DataType::Int64 => {
-                    let arr = result_array
-                        .as_any()
-                        .downcast_ref::<Int64Array>()
-                        .ok_or_else(|| {
-                            ExecutorError::Other("Failed to downcast Int64Array".to_string())
-                        })?;
-                    let max_val = max(arr).ok_or_else(|| {
-                        ExecutorError::Other("SIMD max returned None".to_string())
-                    })?;
-                    Ok(SqlValue::Integer(max_val))
-                }
-                arrow::datatypes::DataType::Float64 => {
-                    let arr = result_array
-                        .as_any()
-                        .downcast_ref::<Float64Array>()
-                        .ok_or_else(|| {
-                            ExecutorError::Other("Failed to downcast Float64Array".to_string())
-                        })?;
-                    let max_val = max(arr).ok_or_else(|| {
-                        ExecutorError::Other("SIMD max returned None".to_string())
-                    })?;
-                    Ok(SqlValue::Double(max_val))
-                }
-                _ => Err(ExecutorError::Other("Unsupported array type for MAX".to_string())),
-            }
-        }
-    }
-}
-
-/// Compute an aggregate over an expression directly from a ColumnarBatch (no row conversion)
-///
-/// This is the batch-native path for expression aggregates. Instead of converting
-/// the batch to rows and then evaluating expressions, we:
-/// 1. Evaluate the expression directly on the batch's column arrays using SIMD
-/// 2. Aggregate the resulting array using SIMD operations
-///
-/// This eliminates the ~10-15ms overhead of `batch.to_rows()` for large batches.
-///
-/// # Arguments
-///
-/// * `batch` - The ColumnarBatch to process (typically already filtered)
-/// * `expr` - The expression to evaluate (e.g., `a * b`)
-/// * `op` - The aggregate operation (SUM, AVG, MIN, MAX, COUNT)
-/// * `schema` - Schema for resolving column names
-///
-/// # Returns
-///
-/// The aggregated SqlValue result
-///
-/// # Performance
-///
-/// For a batch with 100K rows:
-/// - Row-based path: ~15ms (to_rows) + ~5ms (eval) = ~20ms
-/// - Batch-native path: ~3ms (SIMD eval + aggregate)
-///
-/// ~6-7x speedup for expression aggregates.
-#[cfg(feature = "simd")]
-pub(super) fn compute_batch_expression_aggregate(
-    batch: &ColumnarBatch,
-    expr: &Expression,
-    op: AggregateOp,
-    schema: &CombinedSchema,
-) -> Result<SqlValue, ExecutorError> {
-    // Empty batch handling
-    if batch.row_count() == 0 {
-        return Ok(match op {
-            AggregateOp::Count => SqlValue::Integer(0),
-            _ => SqlValue::Null,
-        });
-    }
-
-    // Evaluate expression on batch columns using SIMD
-    let result_array = evaluate_batch_expression(batch, expr, schema)?;
-
-    // Aggregate the result array using SIMD
-    aggregate_column_array(&result_array, op)
-}
-
-/// Evaluate an expression directly on ColumnarBatch column arrays
-///
-/// Returns a ColumnArray containing the computed values.
-#[cfg(feature = "simd")]
-fn evaluate_batch_expression(
-    batch: &ColumnarBatch,
-    expr: &Expression,
-    schema: &CombinedSchema,
-) -> Result<ColumnArray, ExecutorError> {
-    match expr {
-        Expression::ColumnRef { table, column } => {
-            // Simple column reference - return the column directly
-            let col_idx = schema.get_column_index(table.as_deref(), column)
-                .ok_or_else(|| ExecutorError::UnsupportedExpression(
-                    format!("Column not found: {}", column)
-                ))?;
-
-            batch.column(col_idx)
-                .cloned()
-                .ok_or_else(|| ExecutorError::Other(format!(
-                    "Column index {} out of bounds in batch", col_idx
-                )))
-        }
-        Expression::Literal(val) => {
-            // Create an array filled with the literal value
-            create_literal_column_array(val, batch.row_count())
-        }
-        Expression::BinaryOp { left, op, right } => {
-            // Recursively evaluate left and right, then apply operation
-            let left_array = evaluate_batch_expression(batch, left, schema)?;
-            let right_array = evaluate_batch_expression(batch, right, schema)?;
-
-            apply_binary_op_to_columns(&left_array, &right_array, op)
-        }
-        _ => Err(ExecutorError::UnsupportedExpression(
-            "Complex expressions not supported in batch-native columnar aggregates".to_string()
-        )),
-    }
-}
-
-/// Create a ColumnArray filled with a literal value
-#[cfg(feature = "simd")]
-fn create_literal_column_array(value: &SqlValue, len: usize) -> Result<ColumnArray, ExecutorError> {
-    match value {
-        SqlValue::Integer(i) | SqlValue::Bigint(i) => {
-            Ok(ColumnArray::Int64(vec![*i; len], None))
-        }
-        SqlValue::Smallint(i) => {
-            Ok(ColumnArray::Int64(vec![*i as i64; len], None))
-        }
-        SqlValue::Float(f) | SqlValue::Real(f) => {
-            Ok(ColumnArray::Float64(vec![*f as f64; len], None))
-        }
-        SqlValue::Double(f) | SqlValue::Numeric(f) => {
-            Ok(ColumnArray::Float64(vec![*f; len], None))
-        }
-        SqlValue::Null => {
-            // Create array of nulls (represented as Float64 with all nulls)
-            Ok(ColumnArray::Float64(vec![0.0; len], Some(vec![true; len])))
-        }
-        _ => Err(ExecutorError::UnsupportedExpression(format!(
-            "Cannot create literal column array for {:?}", value
-        ))),
-    }
-}
-
-/// Apply a binary operation to two column arrays
-///
-/// Supports SIMD-accelerated arithmetic on Int64 and Float64 columns.
-/// Falls back to row-by-row evaluation for Mixed columns.
-#[cfg(feature = "simd")]
-fn apply_binary_op_to_columns(
-    left: &ColumnArray,
-    right: &ColumnArray,
-    op: &vibesql_ast::BinaryOperator,
-) -> Result<ColumnArray, ExecutorError> {
-    use vibesql_ast::BinaryOperator::*;
-
-    // Try to convert Mixed columns to typed columns for SIMD
-    let (left_typed, right_typed) = match (left, right) {
-        // If either column is Mixed, try to extract numeric values
-        (ColumnArray::Mixed(left_vals), ColumnArray::Mixed(right_vals)) => {
-            let left_f64 = try_extract_f64_from_mixed(left_vals)?;
-            let right_f64 = try_extract_f64_from_mixed(right_vals)?;
-            (ColumnArray::Float64(left_f64, None), ColumnArray::Float64(right_f64, None))
-        }
-        (ColumnArray::Mixed(left_vals), other) => {
-            let left_f64 = try_extract_f64_from_mixed(left_vals)?;
-            (ColumnArray::Float64(left_f64, None), other.clone())
-        }
-        (other, ColumnArray::Mixed(right_vals)) => {
-            let right_f64 = try_extract_f64_from_mixed(right_vals)?;
-            (other.clone(), ColumnArray::Float64(right_f64, None))
-        }
-        _ => (left.clone(), right.clone()),
-    };
-
-    match (&left_typed, &right_typed) {
-        // Both Float64 - direct SIMD operations
-        (ColumnArray::Float64(left_vals, left_nulls), ColumnArray::Float64(right_vals, right_nulls)) => {
-            let result = apply_float64_binary_op(left_vals, right_vals, op)?;
-            let nulls = merge_null_bitmaps(left_nulls.as_deref(), right_nulls.as_deref(), left_vals.len());
-            Ok(ColumnArray::Float64(result, nulls))
-        }
-        // Both Int64 - SIMD operations (result type depends on operation)
-        (ColumnArray::Int64(left_vals, left_nulls), ColumnArray::Int64(right_vals, right_nulls)) => {
-            match op {
-                Plus | Minus | Multiply => {
-                    let result = apply_int64_binary_op(left_vals, right_vals, op)?;
-                    let nulls = merge_null_bitmaps(left_nulls.as_deref(), right_nulls.as_deref(), left_vals.len());
-                    Ok(ColumnArray::Int64(result, nulls))
-                }
-                Divide => {
-                    // Division always produces Float64
-                    let left_f64: Vec<f64> = left_vals.iter().map(|&v| v as f64).collect();
-                    let right_f64: Vec<f64> = right_vals.iter().map(|&v| v as f64).collect();
-                    let result = apply_float64_binary_op(&left_f64, &right_f64, op)?;
-                    let nulls = merge_null_bitmaps(left_nulls.as_deref(), right_nulls.as_deref(), left_vals.len());
-                    Ok(ColumnArray::Float64(result, nulls))
-                }
-                _ => Err(ExecutorError::UnsupportedExpression(format!(
-                    "Unsupported binary operator for Int64: {:?}", op
-                ))),
-            }
-        }
-        // Mixed types - cast to Float64
-        (ColumnArray::Int64(left_vals, left_nulls), ColumnArray::Float64(right_vals, right_nulls)) => {
-            let left_f64: Vec<f64> = left_vals.iter().map(|&v| v as f64).collect();
-            let result = apply_float64_binary_op(&left_f64, right_vals, op)?;
-            let nulls = merge_null_bitmaps(left_nulls.as_deref(), right_nulls.as_deref(), left_vals.len());
-            Ok(ColumnArray::Float64(result, nulls))
-        }
-        (ColumnArray::Float64(left_vals, left_nulls), ColumnArray::Int64(right_vals, right_nulls)) => {
-            let right_f64: Vec<f64> = right_vals.iter().map(|&v| v as f64).collect();
-            let result = apply_float64_binary_op(left_vals, &right_f64, op)?;
-            let nulls = merge_null_bitmaps(left_nulls.as_deref(), right_nulls.as_deref(), left_vals.len());
-            Ok(ColumnArray::Float64(result, nulls))
-        }
-        // Fallback for other types - signal to caller to use row-based path
-        _ => Err(ExecutorError::UnsupportedExpression(
-            "Non-numeric columns not supported in batch arithmetic".to_string()
-        )),
-    }
-}
-
-/// Try to extract f64 values from a Mixed column array
-///
-/// Returns an error if any value is non-numeric.
-#[cfg(feature = "simd")]
-fn try_extract_f64_from_mixed(values: &[SqlValue]) -> Result<Vec<f64>, ExecutorError> {
-    values.iter().map(|v| match v {
-        SqlValue::Integer(i) | SqlValue::Bigint(i) => Ok(*i as f64),
-        SqlValue::Smallint(i) => Ok(*i as f64),
-        SqlValue::Float(f) | SqlValue::Real(f) => Ok(*f as f64),
-        SqlValue::Double(f) | SqlValue::Numeric(f) => Ok(*f),
-        SqlValue::Null => Ok(f64::NAN), // NaN will propagate correctly
-        _ => Err(ExecutorError::UnsupportedExpression(
-            "Non-numeric columns not supported in batch arithmetic".to_string()
-        )),
-    }).collect()
-}
-
-/// Apply a binary operation to Float64 arrays using SIMD
-#[cfg(feature = "simd")]
-fn apply_float64_binary_op(
-    left: &[f64],
-    right: &[f64],
-    op: &vibesql_ast::BinaryOperator,
-) -> Result<Vec<f64>, ExecutorError> {
-    use vibesql_ast::BinaryOperator::*;
-
-    if left.len() != right.len() {
-        return Err(ExecutorError::Other(format!(
-            "Array length mismatch: {} vs {}", left.len(), right.len()
-        )));
-    }
-
-    // SIMD-friendly iteration (compiler will auto-vectorize)
-    let result: Vec<f64> = match op {
-        Plus => left.iter().zip(right.iter()).map(|(l, r)| l + r).collect(),
-        Minus => left.iter().zip(right.iter()).map(|(l, r)| l - r).collect(),
-        Multiply => left.iter().zip(right.iter()).map(|(l, r)| l * r).collect(),
-        Divide => left.iter().zip(right.iter()).map(|(l, r)| l / r).collect(),
-        _ => return Err(ExecutorError::UnsupportedExpression(format!(
-            "Unsupported binary operator for Float64: {:?}", op
-        ))),
-    };
-
-    Ok(result)
-}
-
-/// Apply a binary operation to Int64 arrays using SIMD
-#[cfg(feature = "simd")]
-fn apply_int64_binary_op(
-    left: &[i64],
-    right: &[i64],
-    op: &vibesql_ast::BinaryOperator,
-) -> Result<Vec<i64>, ExecutorError> {
-    use vibesql_ast::BinaryOperator::*;
-
-    if left.len() != right.len() {
-        return Err(ExecutorError::Other(format!(
-            "Array length mismatch: {} vs {}", left.len(), right.len()
-        )));
-    }
-
-    // SIMD-friendly iteration (compiler will auto-vectorize)
-    let result: Vec<i64> = match op {
-        Plus => left.iter().zip(right.iter()).map(|(l, r)| l + r).collect(),
-        Minus => left.iter().zip(right.iter()).map(|(l, r)| l - r).collect(),
-        Multiply => left.iter().zip(right.iter()).map(|(l, r)| l * r).collect(),
-        _ => return Err(ExecutorError::UnsupportedExpression(format!(
-            "Unsupported binary operator for Int64: {:?}", op
-        ))),
-    };
-
-    Ok(result)
-}
-
-/// Merge two null bitmaps (OR operation - if either is null, result is null)
-#[cfg(feature = "simd")]
-fn merge_null_bitmaps(
-    left: Option<&[bool]>,
-    right: Option<&[bool]>,
-    _len: usize,
-) -> Option<Vec<bool>> {
-    match (left, right) {
-        (None, None) => None,
-        (Some(l), None) => Some(l.to_vec()),
-        (None, Some(r)) => Some(r.to_vec()),
-        (Some(l), Some(r)) => {
-            let merged: Vec<bool> = l.iter()
-                .zip(r.iter())
-                .map(|(&l_null, &r_null)| l_null || r_null)
-                .collect();
-            if merged.iter().any(|&is_null| is_null) {
-                Some(merged)
-            } else {
-                None
-            }
-        }
-    }
-}
-
-/// Aggregate a ColumnArray using SIMD operations
-#[cfg(feature = "simd")]
-fn aggregate_column_array(
-    array: &ColumnArray,
-    op: AggregateOp,
-) -> Result<SqlValue, ExecutorError> {
-    match array {
-        ColumnArray::Float64(values, nulls) => {
-            aggregate_f64_array(values, nulls.as_deref(), op)
-        }
-        ColumnArray::Int64(values, nulls) => {
-            aggregate_i64_array(values, nulls.as_deref(), op)
-        }
-        _ => Err(ExecutorError::UnsupportedExpression(
-            "Non-numeric columns not supported for aggregation".to_string()
-        )),
-    }
-}
-
-/// Aggregate a Float64 array
-#[cfg(feature = "simd")]
-fn aggregate_f64_array(
-    values: &[f64],
-    nulls: Option<&[bool]>,
-    op: AggregateOp,
-) -> Result<SqlValue, ExecutorError> {
-    // Filter out null values for aggregation
-    let non_null_values: Vec<f64> = if let Some(null_bitmap) = nulls {
-        values.iter()
-            .zip(null_bitmap.iter())
-            .filter(|(_, &is_null)| !is_null)
-            .map(|(&v, _)| v)
-            .collect()
-    } else {
-        values.to_vec()
-    };
-
-    if non_null_values.is_empty() {
-        return Ok(match op {
-            AggregateOp::Count => SqlValue::Integer(0),
-            _ => SqlValue::Null,
-        });
-    }
-
-    match op {
-        AggregateOp::Sum => {
-            let sum: f64 = non_null_values.iter().sum();
-            Ok(SqlValue::Double(sum))
-        }
-        AggregateOp::Count => {
-            Ok(SqlValue::Integer(non_null_values.len() as i64))
-        }
-        AggregateOp::Avg => {
-            let sum: f64 = non_null_values.iter().sum();
-            let count = non_null_values.len() as f64;
-            Ok(SqlValue::Double(sum / count))
-        }
-        AggregateOp::Min => {
-            let min = non_null_values.iter().cloned().fold(f64::INFINITY, f64::min);
-            Ok(SqlValue::Double(min))
-        }
-        AggregateOp::Max => {
-            let max = non_null_values.iter().cloned().fold(f64::NEG_INFINITY, f64::max);
-            Ok(SqlValue::Double(max))
-        }
-    }
-}
-
-/// Aggregate an Int64 array
-#[cfg(feature = "simd")]
-fn aggregate_i64_array(
-    values: &[i64],
-    nulls: Option<&[bool]>,
-    op: AggregateOp,
-) -> Result<SqlValue, ExecutorError> {
-    // Filter out null values for aggregation
-    let non_null_values: Vec<i64> = if let Some(null_bitmap) = nulls {
-        values.iter()
-            .zip(null_bitmap.iter())
-            .filter(|(_, &is_null)| !is_null)
-            .map(|(&v, _)| v)
-            .collect()
-    } else {
-        values.to_vec()
-    };
-
-    if non_null_values.is_empty() {
-        return Ok(match op {
-            AggregateOp::Count => SqlValue::Integer(0),
-            _ => SqlValue::Null,
-        });
-    }
-
-    match op {
-        AggregateOp::Sum => {
-            let sum: i64 = non_null_values.iter().sum();
-            Ok(SqlValue::Integer(sum))
-        }
-        AggregateOp::Count => {
-            Ok(SqlValue::Integer(non_null_values.len() as i64))
-        }
-        AggregateOp::Avg => {
-            let sum: i64 = non_null_values.iter().sum();
-            let count = non_null_values.len() as f64;
-            Ok(SqlValue::Double(sum as f64 / count))
-        }
-        AggregateOp::Min => {
-            let min = *non_null_values.iter().min().unwrap();
-            Ok(SqlValue::Integer(min))
-        }
-        AggregateOp::Max => {
-            let max = *non_null_values.iter().max().unwrap();
-            Ok(SqlValue::Integer(max))
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::schema::CombinedSchema;
-    use vibesql_catalog::{ColumnSchema, TableSchema};
-    use vibesql_types::DataType;
-    use vibesql_ast::BinaryOperator;
-
-    fn make_test_schema() -> CombinedSchema {
-        let schema = TableSchema::new(
-            "test".to_string(),
-            vec![
-                ColumnSchema::new("price".to_string(), DataType::DoublePrecision, false),
-                ColumnSchema::new("discount".to_string(), DataType::DoublePrecision, false),
-                ColumnSchema::new("quantity".to_string(), DataType::Integer, false),
-            ],
-        );
-        CombinedSchema::from_table("test".to_string(), schema)
-    }
-
-    fn make_test_batch() -> ColumnarBatch {
-        // Create a batch with price, discount, and quantity columns
-        let price_col = ColumnArray::Float64(
-            vec![100.0, 200.0, 300.0, 400.0],
-            None,
-        );
-        let discount_col = ColumnArray::Float64(
-            vec![0.1, 0.2, 0.15, 0.05],
-            None,
-        );
-        let quantity_col = ColumnArray::Int64(
-            vec![10, 20, 15, 25],
-            None,
-        );
-
-        ColumnarBatch::from_columns(
-            vec![price_col, discount_col, quantity_col],
-            Some(vec!["price".to_string(), "discount".to_string(), "quantity".to_string()]),
-        ).unwrap()
-    }
-
-    #[test]
-    #[cfg(feature = "simd")]
-    fn test_batch_expression_aggregate_multiply() {
-        let batch = make_test_batch();
-        let schema = make_test_schema();
-
-        // Test SUM(price * discount)
-        let expr = Expression::BinaryOp {
-            left: Box::new(Expression::ColumnRef {
-                table: None,
-                column: "price".to_string(),
-            }),
-            op: BinaryOperator::Multiply,
-            right: Box::new(Expression::ColumnRef {
-                table: None,
-                column: "discount".to_string(),
-            }),
-        };
-
-        let result = compute_batch_expression_aggregate(&batch, &expr, AggregateOp::Sum, &schema)
-            .expect("Should compute batch expression aggregate");
-
-        // Expected: 100*0.1 + 200*0.2 + 300*0.15 + 400*0.05 = 10 + 40 + 45 + 20 = 115
-        match result {
-            SqlValue::Double(sum) => {
-                assert!((sum - 115.0).abs() < 0.001, "Expected 115.0, got {}", sum);
-            }
-            other => panic!("Expected Double, got {:?}", other),
-        }
-    }
-
-    #[test]
-    #[cfg(feature = "simd")]
-    fn test_batch_expression_aggregate_avg() {
-        let batch = make_test_batch();
-        let schema = make_test_schema();
-
-        // Test AVG(price * discount)
-        let expr = Expression::BinaryOp {
-            left: Box::new(Expression::ColumnRef {
-                table: None,
-                column: "price".to_string(),
-            }),
-            op: BinaryOperator::Multiply,
-            right: Box::new(Expression::ColumnRef {
-                table: None,
-                column: "discount".to_string(),
-            }),
-        };
-
-        let result = compute_batch_expression_aggregate(&batch, &expr, AggregateOp::Avg, &schema)
-            .expect("Should compute batch expression aggregate");
-
-        // Expected: 115.0 / 4 = 28.75
-        match result {
-            SqlValue::Double(avg) => {
-                assert!((avg - 28.75).abs() < 0.001, "Expected 28.75, got {}", avg);
-            }
-            other => panic!("Expected Double, got {:?}", other),
-        }
-    }
-
-    #[test]
-    #[cfg(feature = "simd")]
-    fn test_batch_expression_aggregate_mixed_types() {
-        let batch = make_test_batch();
-        let schema = make_test_schema();
-
-        // Test SUM(price * quantity) - Float64 * Int64
-        let expr = Expression::BinaryOp {
-            left: Box::new(Expression::ColumnRef {
-                table: None,
-                column: "price".to_string(),
-            }),
-            op: BinaryOperator::Multiply,
-            right: Box::new(Expression::ColumnRef {
-                table: None,
-                column: "quantity".to_string(),
-            }),
-        };
-
-        let result = compute_batch_expression_aggregate(&batch, &expr, AggregateOp::Sum, &schema)
-            .expect("Should compute batch expression aggregate");
-
-        // Expected: 100*10 + 200*20 + 300*15 + 400*25 = 1000 + 4000 + 4500 + 10000 = 19500
-        match result {
-            SqlValue::Double(sum) => {
-                assert!((sum - 19500.0).abs() < 0.001, "Expected 19500.0, got {}", sum);
-            }
-            other => panic!("Expected Double, got {:?}", other),
-        }
-    }
-
-    #[test]
-    #[cfg(feature = "simd")]
-    fn test_batch_expression_aggregate_with_literal() {
-        let batch = make_test_batch();
-        let schema = make_test_schema();
-
-        // Test SUM(price * 2)
-        let expr = Expression::BinaryOp {
-            left: Box::new(Expression::ColumnRef {
-                table: None,
-                column: "price".to_string(),
-            }),
-            op: BinaryOperator::Multiply,
-            right: Box::new(Expression::Literal(SqlValue::Integer(2))),
-        };
-
-        let result = compute_batch_expression_aggregate(&batch, &expr, AggregateOp::Sum, &schema)
-            .expect("Should compute batch expression aggregate");
-
-        // Expected: (100 + 200 + 300 + 400) * 2 = 2000
-        match result {
-            SqlValue::Double(sum) => {
-                assert!((sum - 2000.0).abs() < 0.001, "Expected 2000.0, got {}", sum);
-            }
-            other => panic!("Expected Double, got {:?}", other),
-        }
-    }
-
-    #[test]
-    #[cfg(feature = "simd")]
-    fn test_batch_expression_aggregate_nested() {
-        let batch = make_test_batch();
-        let schema = make_test_schema();
-
-        // Test SUM(price * (1 - discount))
-        let expr = Expression::BinaryOp {
-            left: Box::new(Expression::ColumnRef {
-                table: None,
-                column: "price".to_string(),
-            }),
-            op: BinaryOperator::Multiply,
-            right: Box::new(Expression::BinaryOp {
-                left: Box::new(Expression::Literal(SqlValue::Double(1.0))),
-                op: BinaryOperator::Minus,
-                right: Box::new(Expression::ColumnRef {
-                    table: None,
-                    column: "discount".to_string(),
-                }),
-            }),
-        };
-
-        let result = compute_batch_expression_aggregate(&batch, &expr, AggregateOp::Sum, &schema)
-            .expect("Should compute batch expression aggregate");
-
-        // Expected: 100*0.9 + 200*0.8 + 300*0.85 + 400*0.95 = 90 + 160 + 255 + 380 = 885
-        match result {
-            SqlValue::Double(sum) => {
-                assert!((sum - 885.0).abs() < 0.001, "Expected 885.0, got {}", sum);
-            }
-            other => panic!("Expected Double, got {:?}", other),
-        }
-    }
-
-    #[test]
-    #[cfg(feature = "simd")]
-    fn test_batch_expression_aggregate_min_max() {
-        let batch = make_test_batch();
-        let schema = make_test_schema();
-
-        // Test MIN(price * discount)
-        let expr = Expression::BinaryOp {
-            left: Box::new(Expression::ColumnRef {
-                table: None,
-                column: "price".to_string(),
-            }),
-            op: BinaryOperator::Multiply,
-            right: Box::new(Expression::ColumnRef {
-                table: None,
-                column: "discount".to_string(),
-            }),
-        };
-
-        let min_result = compute_batch_expression_aggregate(&batch, &expr, AggregateOp::Min, &schema)
-            .expect("Should compute MIN");
-        let max_result = compute_batch_expression_aggregate(&batch, &expr, AggregateOp::Max, &schema)
-            .expect("Should compute MAX");
-
-        // Values: 10, 40, 45, 20
-        // Min: 10, Max: 45
-        match min_result {
-            SqlValue::Double(min) => {
-                assert!((min - 10.0).abs() < 0.001, "Expected MIN 10.0, got {}", min);
-            }
-            other => panic!("Expected Double for MIN, got {:?}", other),
-        }
-
-        match max_result {
-            SqlValue::Double(max) => {
-                assert!((max - 45.0).abs() < 0.001, "Expected MAX 45.0, got {}", max);
-            }
-            other => panic!("Expected Double for MAX, got {:?}", other),
-        }
-    }
-
-    #[test]
-    #[cfg(feature = "simd")]
-    fn test_batch_expression_aggregate_empty_batch() {
-        let batch = ColumnarBatch::from_columns(
-            vec![
-                ColumnArray::Float64(vec![], None),
-                ColumnArray::Float64(vec![], None),
-            ],
-            Some(vec!["price".to_string(), "discount".to_string()]),
-        ).unwrap();
-
-        let schema = TableSchema::new(
-            "test".to_string(),
-            vec![
-                ColumnSchema::new("price".to_string(), DataType::DoublePrecision, false),
-                ColumnSchema::new("discount".to_string(), DataType::DoublePrecision, false),
-            ],
-        );
-        let combined_schema = CombinedSchema::from_table("test".to_string(), schema);
-
-        let expr = Expression::BinaryOp {
-            left: Box::new(Expression::ColumnRef {
-                table: None,
-                column: "price".to_string(),
-            }),
-            op: BinaryOperator::Multiply,
-            right: Box::new(Expression::ColumnRef {
-                table: None,
-                column: "discount".to_string(),
-            }),
-        };
-
-        // SUM of empty batch should return NULL
-        let sum_result = compute_batch_expression_aggregate(&batch, &expr, AggregateOp::Sum, &combined_schema)
-            .expect("Should handle empty batch");
-        assert_eq!(sum_result, SqlValue::Null);
-
-        // COUNT of empty batch should return 0
-        let count_result = compute_batch_expression_aggregate(&batch, &expr, AggregateOp::Count, &combined_schema)
-            .expect("Should handle empty batch");
-        assert_eq!(count_result, SqlValue::Integer(0));
-    }
-
-    #[test]
-    #[cfg(feature = "simd")]
-    fn test_batch_expression_aggregate_with_nulls() {
-        // Create a batch with some NULL values
-        let price_col = ColumnArray::Float64(
-            vec![100.0, 200.0, 300.0, 400.0],
-            Some(vec![false, true, false, false]), // Second value is NULL
-        );
-        let discount_col = ColumnArray::Float64(
-            vec![0.1, 0.2, 0.15, 0.05],
-            None,
-        );
-
-        let batch = ColumnarBatch::from_columns(
-            vec![price_col, discount_col],
-            Some(vec!["price".to_string(), "discount".to_string()]),
-        ).unwrap();
-
-        let schema = TableSchema::new(
-            "test".to_string(),
-            vec![
-                ColumnSchema::new("price".to_string(), DataType::DoublePrecision, false),
-                ColumnSchema::new("discount".to_string(), DataType::DoublePrecision, false),
-            ],
-        );
-        let combined_schema = CombinedSchema::from_table("test".to_string(), schema);
-
-        let expr = Expression::BinaryOp {
-            left: Box::new(Expression::ColumnRef {
-                table: None,
-                column: "price".to_string(),
-            }),
-            op: BinaryOperator::Multiply,
-            right: Box::new(Expression::ColumnRef {
-                table: None,
-                column: "discount".to_string(),
-            }),
-        };
-
-        let result = compute_batch_expression_aggregate(&batch, &expr, AggregateOp::Sum, &combined_schema)
-            .expect("Should handle batch with nulls");
-
-        // Expected: 100*0.1 + 300*0.15 + 400*0.05 = 10 + 45 + 20 = 75 (skipping NULL row)
-        match result {
-            SqlValue::Double(sum) => {
-                assert!((sum - 75.0).abs() < 0.001, "Expected 75.0, got {}", sum);
-            }
-            other => panic!("Expected Double, got {:?}", other),
-        }
-    }
-}
